+++ src/ws/stubs.rs
@@ -1,10 +1,13 @@
 use std::{
+    cell::RefCell,
     cmp::Ordering,
     collections::HashMap,
+    iter::from_fn,
     sync::{Arc, RwLock},
     time::{Duration, Instant},
 };
 
+use rand::Rng;
 use serde_json::Value;
 use tungstenite::{Bytes, Message, Utf8Bytes};
 
@@ -14,6 +17,7 @@ use crate::matchers::{Body, BodyMatcher, TextMatcher};
 pub struct StubsHandle {
     on_connect: Arc<RwLock<Vec<Stub>>>,
     on_message: Arc<RwLock<Vec<Stub>>>,
+    on_periodical: Arc<RwLock<Vec<Stub>>>,
 }
 
 impl StubsHandle {
@@ -29,11 +33,23 @@ impl StubsHandle {
                     on_message.push(stub);
                 }
             }
+            Stub::Periodical { .. } => {
+                if let Ok(mut on_periodical) = self.on_periodical.write() {
+                    on_periodical.push(stub);
+                }
+            }
         }
     }
 
     pub(crate) fn on_connect(&self, headers: &HashMap<String, String>) -> Option<Msg> {
         Self::get_message(&self.on_connect, headers, None)
+    }
+
+    pub(crate) fn on_periodical(&self, headers: &HashMap<String, String>) -> Option<Vec<Msg>> {
+        let messages: Vec<Msg> =
+            from_fn(|| Self::get_message(&self.on_periodical, headers, None)).collect();
+
+        (!messages.is_empty()).then_some(messages)
     }
 
     pub(crate) fn on_message(
@@ -68,6 +84,11 @@ impl StubsHandle {
 
 // Stubs
 
+thread_local! {
+    static PERIODICALLY_STUBS_INVOCATION_COUNT: RefCell<HashMap<String, usize>> =
+        RefCell::new(HashMap::new());
+}
+
 pub enum Stub {
     Connect {
         headers: Option<HashMap<String, TextMatcher>>,
@@ -77,6 +98,12 @@ pub enum Stub {
         request: RequestMatcher,
         delay: Delay,
         response: Body,
+    },
+    Periodical {
+        id: String,
+        headers: Option<HashMap<String, TextMatcher>>,
+        delay: Delay,
+        responses: Vec<Body>,
     },
 }
 
@@ -121,31 +148,83 @@ impl Stub {
 
                 score
             }
+            Self::Periodical {
+                id,
+                headers,
+                responses,
+                ..
+            } => {
+                let mut score = 1;
+                if let Some(header_matchers) = headers {
+                    for (k, matcher) in header_matchers.iter() {
+                        let header_score = matcher.score(session_headers.get(k));
+                        if header_score != 0 {
+                            score += header_score;
+                        } else {
+                            return 0;
+                        }
+                    }
+                }
+
+                let is_message_available =
+                    PERIODICALLY_STUBS_INVOCATION_COUNT.with(|invocations| {
+                        let map = invocations.borrow();
+                        map.get(id.as_str())
+                            .is_none_or(|&invocation| invocation < responses.len())
+                    });
+                if !is_message_available {
+                    return 0;
+                }
+
+                score
+            }
         }
     }
 
     pub fn message(&self) -> Msg {
         let available_at = match self {
             Self::Connect { .. } => Instant::now(),
-            Self::Message { delay, .. } => match delay {
+            Self::Message { delay, .. } | Self::Periodical { delay, .. } => match delay {
                 Delay::Fixed(delay) => Instant::now()
                     .checked_add(*delay)
                     .unwrap_or_else(Instant::now),
+
+                Delay::Interval(from, to) => {
+                    let from_as_millis: u64 = from.as_millis().try_into().unwrap_or_default();
+                    let to_as_millis: u64 = to.as_millis().try_into().unwrap_or_default();
+                    Instant::now()
+                        .checked_add(Duration::from_millis(
+                            rand::rng().random_range(from_as_millis..to_as_millis),
+                        ))
+                        .unwrap_or_else(Instant::now)
+                }
             },
         };
-        match self {
-            Self::Connect { response, .. } | Self::Message { response, .. } => match &response {
-                Body::Json(json) => Msg(
-                    Message::Text(Utf8Bytes::from(&Value::from(json).to_string())),
-                    available_at,
-                ),
-                Body::PlainText(text) => {
-                    Msg(Message::Text(Utf8Bytes::from(text.as_str())), available_at)
-                }
-                Body::Binary(binary) => {
-                    Msg(Message::Binary(Bytes::from(binary.clone())), available_at)
-                }
-            },
+        let response = match self {
+            Self::Connect { response, .. } | Self::Message { response, .. } => response,
+            Self::Periodical { id, responses, .. } => {
+                let message_idx = PERIODICALLY_STUBS_INVOCATION_COUNT.with(|invocations| {
+                    let mut map = invocations.borrow_mut();
+                    let current_idx = map.entry(id.to_string()).or_insert(0);
+                    let message_idx = *current_idx;
+                    *current_idx += 1;
+                    message_idx
+                });
+                responses
+                    .get(message_idx)
+                    .expect("Always should exist message")
+            }
+        };
+
+        match response {
+            Body::Json(json) => Msg(
+                Message::Text(Utf8Bytes::from(&Value::from(json).to_string())),
+                available_at,
+            ),
+            Body::PlainText(text) => {
+                Msg(Message::Text(Utf8Bytes::from(text.as_str())), available_at)
+            }
+            Body::Binary(binary) => Msg(Message::Binary(Bytes::from(binary.clone())), available_at),
         }
     }
 }
@@ -157,6 +236,7 @@ pub struct RequestMatcher {
 
 pub enum Delay {
     Fixed(Duration),
+    Interval(Duration, Duration),
 }
 
 #[derive(PartialEq, Eq)]