+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 
+# `0.4.0`
+
+- Add periodical message stub. [@manuelgdlvh]
+- Add fixed interval delay. [@manuelgdlvh]
+
 # `0.3.0`
 
 - Add support for define delay of response messages. [@manuelgdlvh]
@@ -8,7 +13,7 @@
 # `0.2.0`
 
 - Add possibility of define custom matcher implementations and new set of predefined matchers for all types. [@manuelgdlvh]
-- Add unitary tests for matchers.
+- Add unitary tests for matchers. [@manuelgdlvh]
 
 # `0.1.1`